+++ src/WorkflowEngineServiceProvider.php
@@ -10,9 +10,11 @@ use Illuminate\Contracts\Events\Dispatcher;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Support\ServiceProvider;
 use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowAdvanceCommand;
+use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowCancelCommand;
 use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowInstancesCommand;
 use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowListCommand;
 use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowPruneCommand;
+use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowRecoverCommand;
 use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowRetryCommand;
 use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowShowCommand;
 use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowSignalCommand;
@@ -68,8 +70,10 @@ final class WorkflowEngineServiceProvider extends ServiceProvider
                 WorkflowStartCommand::class,
                 WorkflowSignalCommand::class,
                 WorkflowAdvanceCommand::class,
+                WorkflowCancelCommand::class,
                 WorkflowRetryCommand::class,
                 WorkflowTickCommand::class,
+                WorkflowRecoverCommand::class,
                 WorkflowPruneCommand::class,
             ]);
         }